feat : GH-27 Add movies screen, and test        GH-21 Move all android test to test only module
