+++ .github/workflows/Build.yml
@@ -102,7 +102,7 @@ jobs:
           disable-animations: true
           disk-size: 6000M
           heap-size: 600M
-          script: ./gradlew connectedDebugAndroidTest --daemon
+          script: ./gradlew :core:androidTest:connectedDebugAndroidTest --daemon
 
       - name: Upload test reports
         if: always()
