+++ build-logic/convention/src/main/kotlin/AndroidFeatureConventionPlugin.kt
@@ -22,10 +22,13 @@
  * THE SOFTWARE.
  */
 
+import me.jerryokafor.ihenkiri.androidTestImplementation
 import me.jerryokafor.ihenkiri.implementation
+import me.jerryokafor.ihenkiri.testImplementation
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.dependencies
+import org.gradle.kotlin.dsl.kotlin
 import org.gradle.kotlin.dsl.project
 
 /*
@@ -68,6 +71,9 @@ class AndroidFeatureConventionPlugin : Plugin<Project> {
                 implementation(project(":core:network"))
                 implementation(project(":core:domain"))
                 implementation(project(":core:model"))
+
+                testImplementation(kotlin("test"))
+                androidTestImplementation(kotlin("test"))
             }
         }
     }
