+++ build-logic/convention/src/main/kotlin/me/jerryokafor/ihenkiri/Jacoco.kt
@@ -101,12 +101,6 @@ internal fun Project.configureJacoco(
                 enableUnitTestCoverage = true
                 enableAndroidTestCoverage = true
             }
-
-            getByName("release") {
-                // disable coverage report on release build
-                enableUnitTestCoverage = false
-                enableAndroidTestCoverage = false
-            }
         }
     }
